extract version utils from update.ts; utils.ts -> version.ts
